requirements.txt
astroid==2.15.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dill==0.3.6
Flask==2.2.3
Flask-HTTPAuth==4.8.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
platformdirs==3.2.0
pypng==0.20220715.0
PyQRCode==1.2.1
python-dotenv==1.0.0
requests==2.29.0
SQLAlchemy==2.0.12
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.2
wrapt==1.15.0
